+++ src/asphalt/serialization/serializers/pickle.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 import pickle
 
-from typeguard import check_argument_types
-
 from asphalt.serialization.api import Serializer
 
 
@@ -20,7 +18,6 @@ class PickleSerializer(Serializer):
     __slots__ = "protocol"
 
     def __init__(self, protocol: int = pickle.HIGHEST_PROTOCOL):
-        assert check_argument_types()
         assert (
             0 <= protocol <= pickle.HIGHEST_PROTOCOL
         ), f'"protocol" must be between 0 and {pickle.HIGHEST_PROTOCOL}'