+++ src/asphalt/serialization/serializers/yaml.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 from io import StringIO
 
 from ruamel.yaml import YAML
-from typeguard import check_argument_types
 
 from asphalt.serialization.api import Serializer
 
@@ -31,7 +30,6 @@ class YAMLSerializer(Serializer):
     __slots__ = "_yaml", "_dumper_options"
 
     def __init__(self, safe: bool = True):
-        assert check_argument_types()
         self._yaml = YAML(typ="safe" if safe else "unsafe")
 
     def serialize(self, obj) -> bytes: